apache-kafka/confluent_cloud_version/requirements.txt
confluent-kafka
python-dotenv
fastavro